Update timestamp field in message payload to be an int

+++ lib/src/main/java/uk/gov/di/ipv/core/library/auditing/AuditEvent.java
@@ -8,14 +8,14 @@ import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport
 @ExcludeFromGeneratedCoverageReport
 @Getter
 public class AuditEvent {
-    @JsonProperty private long timestamp;
+    @JsonProperty private int timestamp;
 
     @JsonProperty("event_name")
     private AuditEventTypes event;
 
     @JsonCreator
     public AuditEvent(
-            @JsonProperty(value = "timestamp", required = true) long timestamp,
+            @JsonProperty(value = "timestamp", required = true) int timestamp,
             @JsonProperty(value = "event_name", required = true) AuditEventTypes event) {
         this.timestamp = timestamp;
         this.event = event;
